Move bulk validation compat bump/email work to a task to prevent request timeout during submission.
